+++ tests/testthat/test-stderr.R
@@ -11,16 +11,18 @@ test_that("Standard errors for RR(receptorLow) are the same", {
 
   rr_glm        <- riskratio(formula = death ~ receptor, data = dat,
                              approach = "glm")
-  rr_glm_start  <- riskratio(formula = death ~ receptor, data = dat,
-                             approach = "glm_start")
+  rr_glm_startp  <- riskratio(formula = death ~ receptor, data = dat,
+                             approach = "glm_startp")
+  rr_glm_startd  <- riskratio(formula = death ~ receptor, data = dat,
+                              approach = "glm_startd")
   rr_robpoisson <- riskratio(formula = death ~ receptor, data = dat,
                              approach = "robpoisson")
   rr_duplicate  <- riskratio(formula = death ~ receptor, data = dat,
                              approach = "duplicate")
   rr_cem        <- riskratio(formula = death ~ receptor, data = dat,
                              approach = "glm_cem")
-  rr_cem_start  <- riskratio(formula = death ~ receptor, data = dat,
-                             approach = "glm_cem_start")
+  rr_cem_startp  <- riskratio(formula = death ~ receptor, data = dat,
+                             approach = "glm_cem_startp")
   rr_margstd    <- riskratio(formula = death ~ receptor, data = dat,
                              approach = "margstd")
   rr_mh <- rr_rd_mantel_haenszel(data = dat, exposure = receptor,
@@ -30,10 +32,10 @@ test_that("Standard errors for RR(receptorLow) are the same", {
   se_glm <- summary(rr_glm)$coefficients["receptorLow", "Std. Error"]
   tol <- 0.01
 
-  expect_equal(se_glm, summary(rr_glm_start)$coefficients["receptorLow",
+  expect_equal(se_glm, summary(rr_glm_startp)$coefficients["receptorLow",
                                                           "Std. Error"],
                tolerance = tol)
-  expect_equal(se_glm, summary(rr_glm_start)$coefficients["receptorLow",
+  expect_equal(se_glm, summary(rr_glm_startd)$coefficients["receptorLow",
                                                           "Std. Error"],
                tolerance = tol)
   expect_equal(se_glm, summary(rr_robpoisson)$coefficients["receptorLow",
@@ -45,7 +47,7 @@ test_that("Standard errors for RR(receptorLow) are the same", {
   expect_equal(se_glm, summary(rr_cem)$coefficients["receptorLow",
                                                     "Std. Error"],
                tolerance = tol)
-  expect_equal(se_glm, summary(rr_cem_start)$coefficients["receptorLow",
+  expect_equal(se_glm, summary(rr_cem_startp)$coefficients["receptorLow",
                                                           "Std. Error"],
                tolerance = tol)
   expect_equal(se_glm, summary(rr_margstd,
@@ -68,14 +70,14 @@ test_that("Standard errors for RD(receptorLow) are the same", {
 
   rd_glm        <- riskdiff(formula = death ~ receptor, data = dat,
                             approach = "glm")
-  rd_glm_start  <- riskdiff(formula = death ~ receptor, data = dat,
-                            approach = "glm_start")
+  rd_glm_startp <- riskdiff(formula = death ~ receptor, data = dat,
+                            approach = "glm_startp")
   rd_robpoisson <- riskdiff(formula = death ~ receptor, data = dat,
                             approach = "robpoisson")
   rd_cem        <- riskdiff(formula = death ~ receptor, data = dat,
                             approach = "glm_cem")
-  rd_cem_start  <- riskdiff(formula = death ~ receptor, data = dat,
-                            approach = "glm_cem_start")
+  rd_cem_startp <- riskdiff(formula = death ~ receptor, data = dat,
+                            approach = "glm_cem_startp")
   rd_margstd    <- riskdiff(formula = death ~ receptor, data = dat,
                             approach = "margstd")
   sum_margstd <- summary(rd_margstd, bootrepeats = 500)
@@ -86,10 +88,7 @@ test_that("Standard errors for RD(receptorLow) are the same", {
   se_glm <- summary(rd_glm)$coefficients["receptorLow", "Std. Error"]
   tol <- 0.01
 
-  expect_equal(se_glm, summary(rd_glm_start)$coefficients["receptorLow",
-                                                          "Std. Error"],
-               tolerance = tol)
-  expect_equal(se_glm, summary(rd_glm_start)$coefficients["receptorLow",
+  expect_equal(se_glm, summary(rd_glm_startp)$coefficients["receptorLow",
                                                           "Std. Error"],
                tolerance = tol)
   expect_equal(se_glm, summary(rd_robpoisson)$coefficients["receptorLow",
@@ -98,7 +97,7 @@ test_that("Standard errors for RD(receptorLow) are the same", {
   expect_equal(se_glm, summary(rd_cem)$coefficients["receptorLow",
                                                     "Std. Error"],
                tolerance = tol)
-  expect_equal(se_glm, summary(rd_cem_start)$coefficients["receptorLow",
+  expect_equal(se_glm, summary(rd_cem_startp)$coefficients["receptorLow",
                                                           "Std. Error"],
                tolerance = tol)
   expect_equal(se_glm, sum_margstd$coefficients["receptorLow", "Std. Error"],